backend/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
PyMongo==4.6.0
dnspython==2.6.1
bcrypt==4.1.2
email-validator==2.1.0